refactor: update option keys in STT and TTS configurations for consistency
- Changed `inactivity_timeout` type from int to float in AssemblyaiOptions to represent seconds.
- Renamed keys in RimeOptions for consistency: `reduceLatency` to `reduce_latency`, `pauseBetweenBrackets` to `pause_between_brackets`, `phonemizeBetweenBrackets` to `phonemize_between_brackets`, `inlineSpeedAlpha` to `inline_speed_alpha`, and `noTextNormalization` to `no_text_normalization`.

+++ livekit-agents/livekit/agents/inference/stt.py
@@ -104,7 +104,7 @@ class AssemblyaiOptions(TypedDict, total=False):
     max_turn_silence: int  # default: not specified
     keyterms_prompt: list[str]  # default: not specified
     language_detection: bool
-    inactivity_timeout: int  # milliseconds
+    inactivity_timeout: float  # seconds
     prompt: str  # default: not specified (u3-rt-pro only, mutually exclusive with keyterms_prompt)
